requirements.txt
Flask==2.3.2 
google-generativeai==0.7.2 
matplotlib==3.7.1 
reportlab==4.2.2
markdown==3.4.3
python-dotenv>=1.0.1
